Allow tasks with no container
Tasks no longer require a container. If the Container is the empty
string the job will be run on the host not inside a container.

Allow starting a Queue without calling InitConfig: default config
applies.

It is an error to add tasks with no defined outputs. This is now checked
and the program will exit if such a job is found.