Shutdown gracefully on SIGINT
Previously on interruption, the code would attempt to kill any running
jobs, but it would still keep submitting and checking them. This led to
race conditions where a job would be kill and then the code would
attempt to check its status causing an error and the program to crash
before all jobs could be gracefully killed. Now on interrpution, all
work is suspended before the jobs are shutdown.